refactor(ai): extract configured task composition (#13875)
