+++ ai/daemons/orchestrator/Orchestrator.mjs
@@ -1,43 +1,39 @@
 // Class bootstrap belongs to `daemon.mjs`; this consumed class relies on global Neo.
-import fs               from 'fs-extra';
-import {spawn}          from 'child_process';
-import net              from 'net';
-import path             from 'path';
-import Base             from '../../../src/core/Base.mjs';
-import ClassSystemUtil  from '../../../src/util/ClassSystem.mjs';
-import AiConfig         from '../../config.mjs';
-import neuralLinkConfig from '../../mcp/server/neural-link/config.mjs';
-import {
-    buildLmsPreloadConfig,
-    buildOllamaReadinessConfig
-} from '../../services/graph/providerReadinessHelper.mjs';
-import HealthService from '../../services/memory-core/HealthService.mjs';
-import SQLite        from '../../graph/storage/SQLite.mjs';
+import fs              from 'fs-extra';
+import {spawn}         from 'child_process';
+import net             from 'net';
+import path            from 'path';
+import Base            from '../../../src/core/Base.mjs';
+import ClassSystemUtil from '../../../src/util/ClassSystem.mjs';
+import AiConfig        from '../../config.mjs';
+import HealthService   from '../../services/memory-core/HealthService.mjs';
+import SQLite          from '../../graph/storage/SQLite.mjs';
 import MaintenanceBackpressureService, {
     DEFAULT_HEAVY_MAINTENANCE_TASK_NAMES,
     DEFAULT_GOLDEN_PATH_DEPENDENCY_TASK_NAMES
 } from './services/MaintenanceBackpressureService.mjs';
-import PrimaryRepoSyncService                                     from './services/PrimaryRepoSyncService.mjs';
-import TenantRepoSyncService                                      from './services/TenantRepoSyncService.mjs';
-import {getDueTask as summaryGetDueTaskImport}                    from './scheduling/summary.mjs';
-import {getDueTask as backupGetDueTaskImport}                     from './scheduling/backup.mjs';
-import {getDueTask as graphLogCompactionGetDueTaskImport}         from './scheduling/graphLogCompaction.mjs';
-import {getDueTask as primaryDevSyncGetDueTaskImport}             from './scheduling/primaryDevSync.mjs';
-import {getDueTask as goldenPathGetDueTaskImport}                 from './scheduling/goldenPath.mjs';
-import {getDueTask as dreamGetDueTaskImport}                      from './scheduling/dream.mjs';
-import {getDueTask as embedDrainLivenessWatchdogGetDueTaskImport} from './scheduling/embedDrainLivenessWatchdog.mjs';
-import memoryCoreConfig                                           from '../../mcp/server/memory-core/config.mjs';
-import MailboxService                                             from '../../services/memory-core/MailboxService.mjs';
-import WakeSubscriptionService                                    from '../../services/memory-core/WakeSubscriptionService.mjs';
-import RequestContextService                                      from '../../mcp/server/shared/services/RequestContextService.mjs';
-import {normalizeAgentIdentityNodeId}                             from '../../scripts/lifecycle/resumeHarness.mjs';
-import TaskStateService                                           from './services/TaskStateService.mjs';
-import ProcessSupervisorService                                   from './services/ProcessSupervisorService.mjs';
-import DreamService                                               from './services/DreamService.mjs';
-import SwarmHeartbeatService                                      from './services/SwarmHeartbeatService.mjs';
-import GoldenPathSynthesizer                                      from '../../services/graph/GoldenPathSynthesizer.mjs';
-import {getDueTask as tenantRepoSyncGetDueTaskImport}             from './scheduling/tenantRepoSync.mjs';
-import {TASK_REGISTRY}                                            from './scheduling/registry.mjs';
+import {buildConfiguredTaskDefinitions as buildConfiguredTaskDefinitionsImport} from './services/ConfiguredTaskDefinitionsService.mjs';
+import PrimaryRepoSyncService                                                   from './services/PrimaryRepoSyncService.mjs';
+import TenantRepoSyncService                                                    from './services/TenantRepoSyncService.mjs';
+import {getDueTask as summaryGetDueTaskImport}                                  from './scheduling/summary.mjs';
+import {getDueTask as backupGetDueTaskImport}                                   from './scheduling/backup.mjs';
+import {getDueTask as graphLogCompactionGetDueTaskImport}                       from './scheduling/graphLogCompaction.mjs';
+import {getDueTask as primaryDevSyncGetDueTaskImport}                           from './scheduling/primaryDevSync.mjs';
+import {getDueTask as goldenPathGetDueTaskImport}                               from './scheduling/goldenPath.mjs';
+import {getDueTask as dreamGetDueTaskImport}                                    from './scheduling/dream.mjs';
+import {getDueTask as embedDrainLivenessWatchdogGetDueTaskImport}               from './scheduling/embedDrainLivenessWatchdog.mjs';
+import memoryCoreConfig                                                         from '../../mcp/server/memory-core/config.mjs';
+import MailboxService                                                           from '../../services/memory-core/MailboxService.mjs';
+import WakeSubscriptionService                                                  from '../../services/memory-core/WakeSubscriptionService.mjs';
+import RequestContextService                                                    from '../../mcp/server/shared/services/RequestContextService.mjs';
+import {normalizeAgentIdentityNodeId}                                           from '../../scripts/lifecycle/resumeHarness.mjs';
+import TaskStateService                                                         from './services/TaskStateService.mjs';
+import ProcessSupervisorService                                                 from './services/ProcessSupervisorService.mjs';
+import DreamService                                                             from './services/DreamService.mjs';
+import SwarmHeartbeatService                                                    from './services/SwarmHeartbeatService.mjs';
+import GoldenPathSynthesizer                                                    from '../../services/graph/GoldenPathSynthesizer.mjs';
+import {getDueTask as tenantRepoSyncGetDueTaskImport}                           from './scheduling/tenantRepoSync.mjs';
+import {TASK_REGISTRY}                                                          from './scheduling/registry.mjs';
 import {
     buildOrchestratorSchedulingOptions,
     runSchedulingPipeline
@@ -45,11 +41,7 @@ import {
 import {
     DEFAULT_DB_PATH,
     DEFAULT_DATA_DIR,
-    DEFAULT_SCRIPT_DIR,
-    buildOllamaServeEnv,
-    getMaxOllamaContextLength,
-    resolveOllamaHostPort,
-    buildTaskDefinitions
+    DEFAULT_SCRIPT_DIR
 } from './taskDefinitions.mjs';
 
 /** @summary Opens/creates the orchestrator sqlite DB via the shared Memory Core schema bootstrap. */
@@ -135,6 +127,7 @@ export class Orchestrator extends Base {
     dreamGetDueTask          = dreamGetDueTaskImport
     goldenPathGetDueTask     = goldenPathGetDueTaskImport
     embedDrainLivenessWatchdogGetDueTask = embedDrainLivenessWatchdogGetDueTaskImport
+    buildConfiguredTaskDefinitionsService = buildConfiguredTaskDefinitionsImport
 
     isPolling                     = false
     pollHandle                    = null
@@ -350,217 +343,18 @@ export class Orchestrator extends Base {
     get neuralLinkBridgeLivenessTimeoutMs() { return AiConfig.orchestrator.neuralLinkBridge.livenessProbeTimeoutMs; }
 
     /**
-     * @summary Builds the supervised task table from the daemon entrypoint's config authorities.
+     * @summary Delegates config-backed task-table construction to the orchestrator task builder.
      * @param {Object} options
      * @param {String} options.scriptDir Script directory.
      * @param {String} options.nodeBin Node executable.
      * @returns {Object}
      */
     buildConfiguredTaskDefinitions({scriptDir, nodeBin}) {
-        const tasks = buildTaskDefinitions({
+        return this.buildConfiguredTaskDefinitionsService({
             scriptDir,
             nodeBin,
-            chromaPort                       : AiConfig.engines.chroma.port,
-            devServerPort                    : AiConfig.orchestrator.devServer.port,
-            devServerLivenessTimeoutMs       : AiConfig.orchestrator.devServer.livenessProbeTimeoutMs,
-            neuralLinkBridgePort             : neuralLinkConfig.port,
             neuralLinkBridgeLivenessTimeoutMs: this.neuralLinkBridgeLivenessTimeoutMs
         });
-
-        this.applyConfiguredGraphLogCompaction(tasks);
-        this.applyConfiguredMlxTask(tasks, {scriptDir});
-        this.applyConfiguredLmsTask(tasks);
-        this.applyConfiguredOllamaTask(tasks);
-
-        return tasks;
-    }
-
-    /**
-     * @summary Applies the graph-log compaction flags owned by AiConfig.
-     * @param {Object} tasks Task table.
-     * @returns {void}
-     */
-    applyConfiguredGraphLogCompaction(tasks) {
-        if (AiConfig.orchestrator.graphLogCompaction.vacuum) {
-            tasks['graphlog-compaction'].args.push('--vacuum');
-        }
-    }
-
-    /**
-     * @summary Adds the orchestrator-owned MLX server task when enabled.
-     * @param {Object} tasks Task table.
-     * @param {Object} options
-     * @param {String} options.scriptDir Script directory.
-     * @returns {void}
-     */
-    applyConfiguredMlxTask(tasks, {scriptDir}) {
-        if (!AiConfig.orchestrator.mlx.enabled) {
-            return;
-        }
-
-        tasks.mlx = {
-            label  : 'mlx inference',
-            command: path.resolve(scriptDir, '../mcp/server/memory-core/.venv/bin/python'),
-            args   : [
-                '-m',
-                'mlx_lm.server',
-                '--model',
-                AiConfig.orchestrator.mlx.model,
-                '--port',
-                String(AiConfig.orchestrator.mlx.port)
-            ],
-            pidFileName    : 'mlx.pid',
-            expectedCommand: 'mlx_lm.server'
-        };
-    }
-
-    /**
-     * @summary Adds the orchestrator-owned LM Studio server task when enabled.
-     * @param {Object} tasks Task table.
-     * @returns {void}
-     */
-    applyConfiguredLmsTask(tasks) {
-        if (!AiConfig.orchestrator.lms.enabled) {
-            return;
-        }
-
-        const preloadConfig  = buildLmsPreloadConfig(AiConfig),
-              requiredModels = Array.isArray(preloadConfig.models)
-                  ? [...new Set(preloadConfig.models.filter(Boolean))]
-                  : [AiConfig.orchestrator.lms.model].filter(Boolean);
-
-        tasks.lms = {
-            label          : 'lms server (LM Studio CLI)',
-            command        : 'lms',
-            args           : ['server', 'start', '--port', String(AiConfig.orchestrator.lms.port)],
-            pidFileName    : 'lms.pid',
-            expectedCommand: 'lms server',
-            requiredModels,
-            // `lms server start` is fire-and-exit: it wakes the LM Studio service and returns, so
-            // the launched server never matches `expectedCommand` for the supervisor's
-            // process-liveness check (`!state.running` stays permanently true). Liveness is the
-            // HTTP endpoint instead — the supervisor gates the restart on this probe so a healthy
-            // server is a silent no-op rather than a re-spawn loop.
-            livenessProbe  : async () => {
-                const {fetchOpenAiCompatibleModelIds} = await import('../../services/graph/providerReadinessHelper.mjs');
-
-                try {
-                    await fetchOpenAiCompatibleModelIds({
-                        host     : AiConfig.openAiCompatible.host,
-                        timeoutMs: AiConfig.orchestrator.providerReadiness.timeoutMs ?? 2000
-                    });
-                    return true;
-                } catch {
-                    return false;
-                }
-            },
-            postSpawn      : async () => {
-                const {ensureLmsModelsLoaded} = await import('../../services/graph/providerReadinessHelper.mjs');
-
-                if (requiredModels.length === 0) {
-                    return {
-                        ready          : true,
-                        loadedModels   : [],
-                        requiredModels,
-                        availableModels: [],
-                        attempts       : 0,
-                        skipped        : true,
-                        reason         : 'no-openai-compatible-local-roles'
-                    };
-                }
-
-                return ensureLmsModelsLoaded({
-                    host          : AiConfig.openAiCompatible.host,
-                    models        : requiredModels,
-                    contextLengths: preloadConfig.contextLengths,
-                    parallels     : preloadConfig.parallels,
-                    allowPartial  : true,
-                    attempts      : AiConfig.orchestrator.providerReadiness.attempts,
-                    delayMs       : AiConfig.orchestrator.providerReadiness.delayMs,
-                    timeoutMs     : AiConfig.orchestrator.providerReadiness.timeoutMs
-                });
-            }
-        };
-    }
-
-    /**
-     * @summary Adds the orchestrator-owned native Ollama server task when enabled.
-     * @param {Object} tasks Task table.
-     * @returns {void}
-     */
-    applyConfiguredOllamaTask(tasks) {
-        if (!AiConfig.orchestrator.ollama.enabled) {
-            return;
-        }
-
-        const readinessConfig = buildOllamaReadinessConfig(AiConfig),
-              roles           = Array.isArray(readinessConfig.roles)
-                  ? readinessConfig.roles.filter(role => role.model)
-                  : [],
-              requiredModels  = [...new Set(roles.map(role => role.model))];
-
-        if (requiredModels.length === 0) {
-            return;
-        }
-
-        tasks.ollama = {
-            label                  : 'ollama server',
-            command                : 'ollama',
-            args                   : ['serve'],
-            pidFileName            : 'ollama.pid',
-            expectedCommand        : 'ollama serve',
-            requiredModels,
-            singletonPort          : resolveOllamaHostPort(readinessConfig.host),
-            duplicateListenerPolicy: 'defer',
-            env                    : buildOllamaServeEnv({
-                host                 : readinessConfig.host,
-                keepAlive            : readinessConfig.keepAlive,
-                contextLength        : getMaxOllamaContextLength(roles),
-                requireParallelModels: readinessConfig.requireParallelModels
-            }),
-            livenessProbe          : async () => {
-                const {ensureOllamaModelsReady} = await import('../../services/graph/providerReadinessHelper.mjs');
-
-                try {
-                    const result = await ensureOllamaModelsReady({
-                        host                 : readinessConfig.host,
-                        roles,
-                        keepAlive            : readinessConfig.keepAlive,
-                        requireParallelModels: readinessConfig.requireParallelModels,
-                        allowPartial         : true,
-                        attempts             : AiConfig.orchestrator.providerReadiness.attempts,
-                        delayMs              : AiConfig.orchestrator.providerReadiness.delayMs,
-                        timeoutMs            : AiConfig.orchestrator.providerReadiness.timeoutMs
-                    });
-
-                    if (
-                        result.error &&
-                        result.availableModels?.length === 0 &&
-                        result.attemptedModels?.length === 0
-                    ) {
-                        return false;
-                    }
-
-                    return true;
-                } catch {
-                    return false;
-                }
-            },
-            postSpawn              : async () => {
-                const {ensureOllamaModelsReady} = await import('../../services/graph/providerReadinessHelper.mjs');
-
-                return ensureOllamaModelsReady({
-                    host                 : readinessConfig.host,
-                    roles,
-                    keepAlive            : readinessConfig.keepAlive,
-                    requireParallelModels: readinessConfig.requireParallelModels,
-                    allowPartial         : true,
-                    attempts             : AiConfig.orchestrator.providerReadiness.attempts,
-                    delayMs              : AiConfig.orchestrator.providerReadiness.delayMs,
-                    timeoutMs            : AiConfig.orchestrator.providerReadiness.timeoutMs
-                });
-            }
-        };
     }
 
     /** @summary Starts the orchestrator timer loop after the wrapper selects this process. */
