+++ ai/deploy/docker-compose.yml
@@ -162,8 +162,8 @@ services:
   # The Agent OS maintenance orchestrator (ADR 0014 §2.3). Built from the shared
   # image via the generalized `SERVICE_ENTRYPOINT` arg. `NEO_AI_DEPLOYMENT_MODE=cloud`
   # activates the cloud-safe scheduler profile — only the cloud-deployable lanes
-  # (summary, backup, dream, golden-path) run; the local-only lanes (primary-dev-sync,
-  # kbSync, bridgeDaemon) are disabled via the Sub A #11722 deployment-mode toggles.
+  # (summary, backup, dream, golden-path) run; local-only lanes and local inference
+  # launchers (primary-dev-sync, kbSync, bridgeDaemon, MLX/LMS/Ollama) are disabled.
   # `cloud` profile — started by `docker compose --profile cloud up`.
   orchestrator:
     build:
@@ -179,6 +179,8 @@ services:
       - NEO_ORCHESTRATOR_BRIDGE_DAEMON_ENABLED=false
       - NEO_ORCHESTRATOR_GOLDEN_PATH_REPO_ENRICHMENT_ENABLED=false
       - NEO_ORCHESTRATOR_MLX_ENABLED=false
+      - NEO_ORCHESTRATOR_LMS_ENABLED=false
+      - NEO_ORCHESTRATOR_OLLAMA_ENABLED=false
       - NEO_TENANT_REPO_MIRROR_ROOT=/app/.neo-ai-data
       - NEO_MODEL_PROVIDER=${NEO_MODEL_PROVIDER:-}
       - NEO_EMBEDDING_PROVIDER=${NEO_EMBEDDING_PROVIDER:-}
